Save alma getirildi.Addmoney butonu eklendi. Asıl sahnenin adı 16 olarak değiştirildi (Pin satırı).14 ve StartMenu sahneleri eklendi.

+++ Assets/Scripts/BallValue.cs
@@ -9,10 +9,10 @@ namespace Pachinko
         public Button ballybtn;
         public CashControl cashcontrol;
         public Text ballText;
-        public double topdeger;
+        public float topdeger;
         public Text ballAtext;
         public CashControl cashControl;
-        public double ballAmount;
+        public float ballAmount;
         public Button ballAA;
         public Button ballAY;
